+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.25] - 2021-07-16
+
+### Added
+- Dependencies updated.
+- Support enumerated and date-time strings in path.
+
+### Fixed
+- Incorrect schema resolution in complex parameters.
+
 ## [0.1.24] - 2021-04-25
 
 ### Added